refactor: ensure tab is reused when opening new one
