+++ package.json
@@ -31,7 +31,7 @@
     "globals": "^13.24.0",
     "monaco-editor": "^0.52.2",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react-dom": "^19.1.0",
     "react-router": "^7.7.0",
     "sparkengineweb": "https://github.com/RuggeroVisintin/SparkEngineWeb/releases/download/sparkengineweb-v0.15.1/sparkengineweb-v0.15.1-lib.tar.gz",
     "styled-components": "^6.1.19",
