+++ src/__mocks__/InMemoryEventBusDouble.tsx
@@ -1,0 +1,19 @@
+import { EventBus } from "../core/common/ports";
+
+export class InMemoryEventBusDouble implements EventBus {
+    private subscribers: { [key: string]: ((event: any) => void)[] } = {};
+
+    subscribe<T>(eventName: string, callback: (event: T) => void): void {
+        if (!this.subscribers[eventName]) {
+            this.subscribers[eventName] = [];
+        }
+
+        this.subscribers[eventName].push(callback);
+    }
+
+    publish<T>(eventName: string, event: T): void {
+        if (this.subscribers[eventName]) {
+            this.subscribers[eventName].forEach(callback => callback(event));
+        }
+    }
+}
