+++ src/core/common/adapters/index.tsx
@@ -1,0 +1,2 @@
+export * from './FileSystemRepository';
+export * from './ReactStateRepository';
