+++ src/core/common/index.tsx
@@ -1,4 +1,4 @@
-export * from './FileSystemRepository';
+export * from './adapters/FileSystemRepository';
 export * from './uuidToRgb';
 export * from './dependencyInjection';
 export * from './utils';
