+++ src/core/common/ports/StateRepository.ts
