+++ src/core/scripting/application/ScriptEditorService.test.tsx
@@ -1,26 +1,8 @@
 import { ReactStateRepository } from "../../editor";
 import { OpenScriptingEditorCommand } from "../domain/commands";
-import { EventBus } from "./ports/EventBus";
 import { ScriptEditorService } from "./ScriptEditorService";
 import { ScriptEditorState } from "./ScriptEditorState";
-
-class InMemoryEventBusDouble implements EventBus {
-    private subscribers: { [key: string]: ((event: any) => void)[] } = {};
-
-    subscribe<T>(eventName: string, callback: (event: T) => void): void {
-        if (!this.subscribers[eventName]) {
-            this.subscribers[eventName] = [];
-        }
-
-        this.subscribers[eventName].push(callback);
-    }
-
-    publish<T>(eventName: string, event: T): void {
-        if (this.subscribers[eventName]) {
-            this.subscribers[eventName].forEach(callback => callback(event));
-        }
-    }
-}
+import { InMemoryEventBusDouble } from "../../../__mocks__/InMemoryEventBusDouble";
 
 describe('core/scripting/application/ScriptEditorService', () => {
     let service: ScriptEditorService;
