+++ src/templates/EntityPropsPanel/index.tsx
@@ -4,9 +4,6 @@ import { FormInput } from "../../components";
 import { Box, Button, Spacing } from "../../primitives";
 import { InputRow } from "../../primitives/InputRow";
 import { MaterialPropsGroup } from "./components/MaterialPropsGroup";
-import { Link } from "react-router";
-import { EventBusWithBrowserBroadcast } from "../../core/scripting/infrastructure";
-import { OpenScriptingEditorCommand } from "../../core/scripting/domain/commands";
 
 interface EntityPropsPanelProps {
     currentEntity?: IEntity,
@@ -78,7 +75,6 @@ const TransformPropsGroup = ({ transform, onUpdateSize, onUpdatePosition }: Tran
 export const EntityPropsPanel = ({ currentEntity, onUpdatePosition, onUpdateSize, onMaterialUpdate }: EntityPropsPanelProps) => {
     const transform = currentEntity?.getComponent<TransformComponent>('TransformComponent');
     const material = currentEntity?.getComponent<MaterialComponent>('MaterialComponent');
-    const eventBus = new EventBusWithBrowserBroadcast('scripting');
 
     return (
         <Box $size={1} $scroll $divide $spacing={Spacing.lg}>
@@ -91,22 +87,21 @@ export const EntityPropsPanel = ({ currentEntity, onUpdatePosition, onUpdateSize
             />}
             {typeOf(currentEntity) === 'TriggerEntity' && (
                 <Box data-testid="EntityPropsPanel.TriggerEntity.ScriptingProp">
-                    <Button>
-                        <Link
-                            to={'/scripting/'}
-                            data-testid="EntityPropsPanel.TriggerEntity.ScriptingLink"
-                            target="_blank"
-                            rel="noopener noreferrer"
-                            onClick={() => eventBus.publish<OpenScriptingEditorCommand>('OpenScriptingEditorCommand', {
-                                currentScript: 'console.log("hello world");',
-                                entityUuid: 'test-entity-uuid'
-                            })}
-                        >
-                            <Box> Open Scripting </Box>
-                        </Link>
+                    <Button
+                        data-testid="EntityPropsPanel.TriggerEntity.ScriptingLink"
+                        onClick={() => {
+                            const namedWindow = window.open('/scripting', 'scripting');
+
+                            if (namedWindow) {
+                                namedWindow.focus();
+                            }
+                        }}
+                    >
+                        <Box> Open Scripting </Box>
                     </Button>
                 </Box>
-            )}
-        </Box>
+            )
+            }
+        </Box >
     )
 }
